Fixed some bugs direct_sum did not output a Matrix instance and changed dtype to float #1 #11

+++ eigenqr/vector.py
@@ -152,4 +152,4 @@ def zero_vector(n):
 
     """
 
-    return np.array(np.zeros((n,), dtype=int))
+    return np.array(np.zeros((n,)))
